+++ game_patch/misc/alpine_settings.cpp
@@ -759,11 +759,10 @@ bool alpine_player_settings_load(rf::Player* player)
         g_alpine_game_config.use_new_savegame_format = std::stoi(settings["UseAlpineSavegameFormat"]);
         processed_keys.insert("UseAlpineSavegameFormat");
     }
-    if (settings.count("SavegameHighAccuracy")) {
-        g_alpine_game_config.savegame_high_accuracy = std::stoi(settings["SavegameHighAccuracy"]);
-        processed_keys.insert("SavegameHighAccuracy");
+    if (settings.count("SpeedrunSavegameMode")) {
+        g_alpine_game_config.speedrun_savegame_mode = std::stoi(settings["SpeedrunSavegameMode"]);
+        processed_keys.insert("SpeedrunSavegameMode");
     }
-    asg::g_use_high_accuracy_savegame = g_alpine_game_config.savegame_high_accuracy;
 
     // Load multiplayer settings
     if (settings.count("MultiplayerCharacter")) {
@@ -1233,7 +1232,7 @@ void alpine_player_settings_save(rf::Player* player)
     file << "StaticBombCode=" << g_alpine_game_config.static_bomb_code << "\n";
     file << "ExposureDamage=" << g_alpine_game_config.apply_exposure_damage << "\n";
     file << "UseAlpineSavegameFormat=" << g_alpine_game_config.use_new_savegame_format << "\n";
-    file << "SavegameHighAccuracy=" << g_alpine_game_config.savegame_high_accuracy << "\n";
+    file << "SpeedrunSavegameMode=" << g_alpine_game_config.speedrun_savegame_mode << "\n";
 
     // Multiplayer
     file << "\n[MultiplayerSettings]\n";
@@ -1299,7 +1298,6 @@ void set_alpine_config_defaults() {
     rf::game_set_gore_level(2);
     rf::g_fast_animations = false;
     g_alpine_game_config.save_console_history = true; // must be set here because evaluated before config loaded
-    asg::g_use_high_accuracy_savegame = g_alpine_game_config.savegame_high_accuracy;
     build_time_left_string_format();
     set_play_sound_events_volume_scale();
     apply_entity_sim_distance();